update install script

+++ scripts/install-boost.bash
@@ -2,6 +2,15 @@
 
 # Install boost version 1.74.0
 
+set -e
+
+function clean_before_exit {
+	rm -rf $current_directory/libs/boost
+	rm -rf $current_directory/.tmp
+}
+
+trap clean_before_exit EXIT
+
 current_directory=$(pwd)
 
 include=libs/boost/include
@@ -9,12 +18,21 @@ bin=libs/boost/bin
 mkdir -p $include
 mkdir -p $bin
 
+cache=$(pwd)/.cache
+
+mkdir -p $cache
+
+boost_archive=$cache/boost.tar.gz
+boost_dl_url=https://dl.bintray.com/boostorg/release/1.74.0/source/boost_1_74_0.tar.gz
+
+if [ ! -f "$boost_archive" ]
+then
+	curl -o $boost_archive -L $boost_dl_url
+fi
+
 mkdir .tmp
 cd .tmp
 
-boost_archive=boost.tar.gz
-boost_dl_url=https://dl.bintray.com/boostorg/release/1.74.0/source/boost_1_74_0.tar.gz
-curl -o $boost_archive -L $boost_dl_url
 tar -xf $boost_archive
 
 cd boost_1_74_0
@@ -24,3 +42,5 @@ cd boost_1_74_0
 
 cd ../..
 rm -rf .tmp
+
+trap - EXIT
